+++ src/Basque/Mews.Fiscalizations.Basque.Tests/Bizkaia/Helpers/XmlSerializationHelper.cs
@@ -1,6 +1,6 @@
 ﻿using Mews.Fiscalizations.Core.Xml;
 
-namespace Mews.Fiscalizations.Basque.Tests.Bizkaia
+namespace Mews.Fiscalizations.Basque.Tests.Bizkaia.Helpers
 {
     public static class XmlSerializationHelper
     {
